pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.7",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "aiosqlite>=0.20.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=9.0.2"]
